Implemented aggregate weights  that take care of: -  domain size - Invalid domain id - invalid id : otherwise the user will pay like he published but without any reward just because somebody was faster... Like say in Naples Horned and Beaten. Come back to u32 for aggregation len.

+++ pallets/aggregate/src/data.rs
@@ -20,7 +20,7 @@ use scale_info::TypeInfo;
 use sp_core::{Get, H256};
 use sp_runtime::{traits::Keccak256, BoundedBTreeMap, BoundedVec};
 
-pub type AggregationSize = u8;
+pub type AggregationSize = u32;
 
 #[derive(Debug, Clone, PartialEq, Encode, Decode, TypeInfo, MaxEncodedLen)]
 pub struct StatementEntry<A, B> {
